+++ app.js
@@ -40,11 +40,7 @@ const errorHandler = (err, req, res, next) => {
     });
   } else {
     // unexpected error, so log it
-    logger.error({
-      error: 'Unexpected error',
-      details: err,
-      endpoint: req.originalUrl,
-    });
+    logger.error(err);
     res.status(500);
     res.json({ errors: ['An unexpected error has occurred.'] });
   }
@@ -66,6 +62,12 @@ app.use(log.addUID, log.logRequestInfo);
 app.get('/', (req, res) => {
   res.redirect('/api-docs');
 });
+
+// everything else is API functions, so set the Content-Type header here 
+app.use((req, res, next) => {
+  res.setHeader('Content-Type', 'application/json');
+  next();
+})
 
 app.get('/available', fileQc.getAvailableConstants);
 
@@ -144,12 +146,8 @@ app.get('/metrics', async (req, res) => {
   try {
     const mostRecentImportTime = await fileQc.getMostRecentFprImportTime();
     prom.mostRecentFprImport.set(mostRecentImportTime);
-  } catch (e) {
-    logger.error({
-      error: 'Error getting most recent File Provenance Report import time',
-      details: e,
-      method: '/metrics endpoint',
-    });
+  } catch (err) {
+    logger.error(err);
   }
   res.set('Content-Type', prom.prometheus.register.contentType);
   res.end(await prom.prometheus.register.metrics());