+++ utils/controllerUtils.js
@@ -73,6 +73,42 @@ const bonusMsg = (entityType, entityIdentifier, whatIsBonus, bonusItems) => {
   return `The request for ${entityType} ${entityIdentifier} contains extra ${whatIsBonus}: (${bonusItems}) compared to those which are present in the existing ${entityType}.`
 }
 
+function streamResponse (req, res, daoFn, transformAndSend, methodName, logger) {
+    const streamed = daoFn((stream) => {
+      res.status(200);
+      req.on('close', () => {
+        // destroy the stream if the request ended early, before the streamed data finished transmission.
+        // destroy by passing an error to trigger pg-promise cleanup (including releasing the database connection)
+        if (!res.writableEnded) {
+          stream.destroy(new Error('Client disconnected; destroying stream'));
+        }
+      });
+      transformAndSend(stream)();
+      stream.on('error', (err) => {
+        if (err.message === 'Client disconnected; destroying stream') {
+          return; // client has disconnected, so no sense in trying to return an HTTP status code
+        }
+        logger.error(err);
+        if (!res.writableEnded) {
+          res.status(500).end();
+        }
+      });
+    });
+    streamed.catch((err) => {
+      if (err.message === 'Client disconnected; destroying stream') {
+        logger.info('Stream destroyed due to client disconnect');
+        return;
+      }
+      // For real errors, log them instead of re-throwing into the void
+      logger.error(err);
+    });
+    logger.info({
+      streamRowsProcessed: streamed.processed,
+      streamingDuration: streamed.duration,
+      method: methodName,
+    });
+}
+
 module.exports = {
   ValidationError: ValidationError,
   ConflictingDataError: ConflictingDataError,
@@ -82,4 +118,5 @@ module.exports = {
   handleErrors: handleErrors,
   missingMsg: missingMsg,
   bonusMsg: bonusMsg,
+  streamResponse: streamResponse,
 };